delete redundant test, rename package
